requirements.txt
python-telegram-bot==20.3
redis==4.5.4
python-dotenv==1.0.0
PyYAML==6.0
apscheduler
requests